refactor: refactor setting page
